added a declaration/definition of the ucc:uccParser::error(location, message) function and everything compiles. Albeit i get a segfault on running.
-- Nothing has been tested
   I don't know if any of the code runs

With that in mind I know that a few things need improvement:

The Symbol Table needs to be tested. It can be optimized a little more as well as it was
only a direct transfer from crap C code from before. As In all of the code.

The List class needs to be tested and cleaned. I'm not sure it works.

The Code Generator needs to be cleaned and tested.