squashing if and while expressions had a few lables misplaced. May have fixed them all, may have introduced 1 new one.
currently looking for a function declaration/definiton bug when the arugment is only void

also trying to figure out why the pushga code is coming out wrong.
The getleveldiff() function /* or something like that*/ which calculates
the difference in scope levels between a symbol and the current scope. It appears to be off
by 1 right now.